[BE] fix: 회원탈퇴(멤버 삭제) 시 데이터 무결성 오류 수정
- 멤버 삭제할 때 해당 멤버가 작성한 피드가 존재할 시 외래키 관련 오류 발생 (a foreign key constraint fails )
- 수정 전 : feed entity에서 @ElementCollection 어노테이션 사용하여 List<ChildFeedRequest> 저장 및 관리
- 수정 후 : @ElementCollection를 @OneToMany로 변경 및 ChildFeedRequest를 entity로 변경 List<ChildFeedRequest> 관리하여 무결성 오류 수정
