--- BE/toonners/src/main/java/com/example/toonners/domain/feed/dto/request/ChildFeedRequest.java
@@ -1,25 +1,0 @@
-package com.example.toonners.domain.feed.dto.request;
-
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-import java.util.List;
-import java.util.Set;
-
-@Setter
-@Getter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Embeddable
-public class ChildFeedRequest {
-    private Float starring;
-    private Set<String> hashtagGenre;
-    private Set<String> hashtagVibe;
-    private String title;
-    private String imageUrl;
-    private String siteUrl;
-    private Float rating;
-    private List<String> days;
-}
