+++ BE/toonners/src/main/java/com/example/toonners/domain/feed/dto/response/ChildFeedResponse.java
@@ -1,6 +1,6 @@
 package com.example.toonners.domain.feed.dto.response;
 
-import com.example.toonners.domain.feed.dto.request.ChildFeedRequest;
+import com.example.toonners.domain.feed.entity.ChildFeedRequest;
 import lombok.*;
 
 import java.util.Set;
