+++ BE/toonners/src/main/java/com/example/toonners/domain/feed/service/FeedService.java
@@ -5,10 +5,10 @@ import com.example.toonners.domain.bookmark.entity.Bookmark;
 import com.example.toonners.domain.bookmark.repository.BookmarkRepository;
 import com.example.toonners.domain.chatRoom.entity.ChatRoom;
 import com.example.toonners.domain.chatRoom.repository.ChatRoomRepository;
-import com.example.toonners.domain.feed.dto.request.ChildFeedRequest;
 import com.example.toonners.domain.feed.dto.request.CreateFeedRequest;
 import com.example.toonners.domain.feed.dto.request.UpdateFeedRequest;
 import com.example.toonners.domain.feed.dto.response.FeedInfoResponse;
+import com.example.toonners.domain.feed.entity.ChildFeedRequest;
 import com.example.toonners.domain.feed.entity.Feed;
 import com.example.toonners.domain.feed.repository.FeedRepository;
 import com.example.toonners.domain.like.repository.LikeRepository;
